fix: кодировка конфига теперь типо-безопасна для передачи по URL. Увеличен лимит итераций хэширования PBKDF2. Валидация уникальности символов при генерации с точным подсчетом символов

+++ lib/core/services/navigation_service.dart
@@ -48,5 +48,5 @@ enum AppTab {
   final String label;
 
   static AppTab fromIndex(int index) =>
-      values[index.clamp(0, values.length - 1)];
+      values[index.clamp(0, values.length - 1).toInt()];
 }
